requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
qdrant-client
pydantic
python-dotenv
GitPython
typer
httpx
python-multipart
deepeval
